requirements.txt
alembic==1.18.4
altair==6.0.0
attrs==26.1.0
bcrypt==5.0.0
black==26.3.1
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
narwhals==2.19.0
numpy==2.2.6
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pathspec==1.0.4
pillow==12.2.0
platformdirs==4.9.4
protobuf==7.34.1
PyYAML==6.0.3
pyarrow==23.0.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
pytz==2026.1.post1
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.49
streamlit==1.56.0
streamlit-authenticator==0.4.2
tenacity==9.1.4
toml==0.10.2
tomli==2.4.1
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
